% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all.R
\name{A.dot}
\alias{A.dot}
\title{Dot file reporesenting a pedigree tree.}
\usage{
A.dot(A, width = NA, dotfile = "A.dot")
}
\arguments{
\item{A}{The numerator relationship matrix.}

\item{width}{Width of the graph node character identifiers. If \code{NA} (the
default) the complete identity names are written to the dot file.}

\item{dotfile}{The output dot file name, default is \code{"dot.dot"}.}
}
\value{
A two column matrix of the graph node relationships.
}
\description{
Exports the relationship structure in a relationship matrix as a
directed graph in a dot file suitable for plotting.
}
\details{
The matrix \code{A} can be decomposed as \code{A=LDU}, where
\code{U=t(L)} and \code{inv(L)} represents the directed graph of
additive genetic relationships in the pedigree. This matrix can be
converted to a graph object in dot file format suitable for
plotting by \code{Graphviz}, say.

The function calls \code{zapsmall} to round nuisance values to
zero; this can be controlled through the \code{"digits"} component
of \code{options()}
}
\section{Side effects}{


The \code{dotfile} is written to the working directory.
}

