% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all.R
\name{ainv}
\alias{ainv}
\title{Calculate an inverse relationship matrix.}
\usage{
ainv(
  pedigree,
  fgen = list(character(0), 0.01),
  gender = character(0),
  groups = 0,
  groupOffset = 0,
  selfing = NA,
  inBreed = NA,
  mgs = FALSE,
  mv = c("NA", "0", "*"),
  psort = FALSE
)
}
\arguments{
\item{pedigree}{A data frame where the first three columns correspond to the
identifiers for the individual, male parent and female parent,
respectively. The row giving the pedigree of an individual must
appear before any row where that individual appears as a
parent. Founders use 0 (zero) or NA in the parental columns.}

\item{fgen}{An optional list of length 2 where \code{fgen[[1]]} is a character
string naming the column in \code{pedigree} that contains the level
of selfing or the level of inbreeding of an individual. In
\code{pedigree[,fgen[[1]]]}, 0 indicates a simple cross, 1
indicates selfed once, 2 indicates selfed twice, etc. A value
between 0 and 1 for a base individual is taken as its inbreeding
value. If the pedigree has implicit individuals (they appear as
parents but not as individuals), they will be assumed base
non-inbred individuals unless their inbreeding level is set with
\code{fgen[[2]]}, where \code{0 < fgen[[2]] < 1} is the inbreeding
level of such individuals.}

\item{gender}{An optional character string naming the column of \code{pedigree}
that codes for the gender of an individual. \code{pedigree[,
gender]} is coerced to a factor and must only have two (arbitrary)
levels, the first of which is taken to mean "male". An inverse
relationship matrix is formed for the X chromosome as described by
\cite{Fernando and Grossman, 1990} for species where the male is XY
and the female is XX.}

\item{groups}{An integer scalar (\emph{g}) indicating genetic groups in the
pedigree. The first \emph{g} lines of the pedigree identify the
genetic groups (with zero in both the male and female parent
columns). All other rows must specify one of the genetic groups as
the male or female parent if the actual parent is unknown. The
default is \eqn{g=0}.}

\item{groupOffset}{A numeric scalar \emph{e > 0} added to the diagonal elements of
\eqn{A^{-1}} pertaining to \code{groups}, shrinking the group
effects by \eqn{e}. When a constant is added, no adjustment of the
degrees of freedom is made for genetic groups.  Set to -1 to add no
offset but to suppress insertion of constraints where empty groups
appear; the empty groups are then not counted in the degress of
freedom adjustment. The default is \eqn{e=0}.}

\item{selfing}{A numeric scalar (\emph{s}) allowing for partial selfing when the
third field of \code{pedigree} is unknown. It indicates that
progeny from a cross where the male parent is unknown is assumed to
be from selfing with probability \emph{s} and from outcrossing with
probability \emph{(1-s)}. This is appropriate in some forestry tree
breeding studies where seed collected from a tree may have been
pollinated by the mother tree or pollinated by some other tree
(\cite{Dutkowski and Gilmour, 2001}). Do not use the \code{selfing}
argument in conjunction with \code{inBreed} or \code{mgs}.}

\item{inBreed}{A numeric scalar (default \code{NA}) giving the inbreeding
coefficient for \emph{base} individuals. This argument generates
the numerator relationship matrix for inbred lines. Each cross is
assumed to be selfed several times to stabilize as an inbred line
as is usual for cereal crops, for example, before being evaluated
or crossed with another line.  Since inbreeding is usually
associated with strong selection, it is not obvious that a pedigree
assumption of covariance of 0.5 between parent and offspring
actually holds. The \code{inBreed} argument cannot be used in
conjunction with \code{selfing} or \code{mgs}.}

\item{mgs}{If \code{TRUE} (default \code{FALSE}), the third identity in the
pedigree is the male parent of the female parent (maternal
grand-sire) rather than the female parent.}

\item{mv}{A character vector of missing value indicators; elements of
\code{pedigree} that exactly match any of the members of \code{mv}
are treated as missing.}

\item{psort}{If \code{TRUE} (default \code{FALSE}), the pedigree data frame is
returned in founder order after any insertions and permutations.}
}
\value{
A three-column matrix with class \code{ginv} holding the lower
triangle of the inverse relationship matrix in sparse form. The
first 2 columns are the \emph{row} and \emph{column} indices,
respectively, and the third column holds the inverse matrix element
itself. Sort order is columns within rows, that is, the lower
triangle row-wise. This matrix has attributes:

\describe{

\item{\code{rowNames}}{A character vector of identifiers for the
rows of the matrix.}

\item{\code{inbreeding}}{A numeric vector containing the
inbreeding coefficient for each individual, calculated as
\code{diag(A-I)}.}

\item{\code{geneticGroups}}{A numeric vector of length 2 containing
the \code{groups} and \code{groupOffset} arguments.}

\item{\code{logdet}}{The log determinant.}

}
}
\description{
Generates an inverse relationship matrix in sparse triplet form
from a pedigree data frame.
}
\details{
Uses the method of \cite{Meuwissen and Luo, 1992} to compute the inverse
relationship matrix directly from the pedigree.
}
\examples{

\dontrun{

# Simple pedigree

ped <- data.frame(me = c(1,2,3,4,5,6,7,8,9,10),
                  dad = c(0,0,0,1,1,2,4,5,7,9),
                  mum = c(0,0,0,1,1,2,6,6,8,9))
p.ai <- ainv(ped)

# Known filial generation

pdfg <- data.frame(me = c(1,2,3,4,5,6,7),
                   dad = c(0,0,1,1,1,1,1),
                   mum = c(0,0,0,2,2,2,2),
                   fgen = c(0.8,0.0,2.0,0.0,2.0,3.0,0.0))
pdfg.ai <- ainv(pdfg,fgen=list('fgen',0.4))
pdfg.mat <- sparse2mat(pdfg.ai)
zapsmall(solve(pdfg.mat))
zapsmall(cbind(pdfg.a$inbreeding,diag(pdfg.mat)))
}

}
\references{
% bibentry: Dutkowski:2001

% end:bibentry:  Dutkowski:2001
% bibentry: Fernando:1990

% end:bibentry:  Fernando:1990
% bibentry: Meuwissen:1992

% end:bibentry:  Meuwissen:1992
}
