% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all.R
\name{AIsweep}
\alias{AIsweep}
\title{Remove individuals from an inverse relationship matrix.}
\usage{
AIsweep(ginv, keep)
}
\arguments{
\item{ginv}{A matrix containing an inverse numerator relationship matrix in
three column co-ordinate form with a \code{rowNames} attribute.}

\item{keep}{A character vector naming the individuals to retain.}
}
\value{
A three column sparse coordinate form matrix in row major order
with attribute \code{"INVERSE"} set to \code{TRUE}.
}
\description{
Deletes individuals from an inverse numerator relationship matrix
by sweeping out the respective rows.
}
