% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all.R
\name{checkPedigree}
\alias{checkPedigree}
\title{Check a pedigree file}
\usage{
checkPedigree(
  pedigree,
  fgen = list(character(0), 0.01),
  gender = character(0),
  mv = c("0", "*", " "),
  verbose = FALSE
)
}
\arguments{
\item{pedigree}{A data frame with (at least) three columns that correspond to the
individual, male parent and female parent, respectively. The row
giving the pedigree of an individual must appear before any row
where that individual appears as a parent. Founders use 0 (zero) or
NA in the parental columns.}

\item{fgen}{An optional list of length 2 where \code{fgen[[1]]} is a character
string naming the column in \code{pedigree} that contains the level
of selfing or the level of inbreeding of an individual. In
\code{pedigree[,fgen[[1]]]}, 0 indicates a simple cross, 1
indicates selfed once, 2 indicates selfed twice, etc. A value
between 0 and 1 for a base individual is taken as its inbreeding
value. If the pedigree has implicit individuals (they appear as
parents but not as individuals), they will be assumed base
non-inbred individuals unless their inbreeding level is set with
\code{fgen[[2]]} where \code{0<fgen[[2]]<1} is the inbreeding level
of such individuals.}

\item{gender}{An optional character string naming the column of \code{pedigree}
that codes for the gender of an individual. \code{pedigree[,
gender]} is coerced to a factor and must only have two (arbitrary)
levels, the first of which is taken to mean "male". An inverse
relationship matrix is formed for the X chromosome as described by
\cite{Fernando and Grossman (1990)} for species where the male is
XY and the female is XX.}

\item{mv}{Missing value indicator; elements of \code{pedigree} that exactly
match any element of \code{mv} are treated as missing.}

\item{verbose}{If \code{TRUE} details of pedigree insertions and relocations are printed;
the default is \code{FALSE} to supress printing.}
}
\value{
A data frame containing the expanded or re-ordered pedigree.
}
\description{
Check a pedigree file for missing or out of order founders.
}
\references{
% bibentry: Fernando:1990

% end:bibentry:  Fernando:1990
}
