% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all.R
\name{count_gen}
\alias{count_gen}
\title{Count generation number.}
\usage{
count_gen(ped)
}
\arguments{
\item{ped}{A data frame with (at least) three columns that correspond to the
individual, male parent and female parent, respectively. The row
giving the pedigree of an individual must appear before any row
where that individual appears as a parent. Founders use 0 (zero) or
\code{NA} in the parental columns.}
}
\value{
A numeric vector of length \code{nrow(ped)} containing the
generation number for each individual.
}
\description{
Counts the generation number for each individual in a pedigreee.
}
