% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all.R
\name{equiv.mm}
\alias{equiv.mm}
\title{Equivalence classes for (near) co-incident genetic markers.}
\usage{
equiv.mm(M, na.match = FALSE, threshold = 0.05)
}
\arguments{
\item{M}{An \code{n x m} matrix of marker scores for m markers on n individuals. The
matrix must be numeric and capable of being coerced to integer
values.}

\item{na.match}{If \code{FALSE} (the default), \code{NA} in either marker counts as
a mismatch, that is \code{NAs} match nothing. If \code{TRUE},
\code{NA} in either marker counts as a match, that is \code{NAs}
match everything.}

\item{threshold}{Marker pairs whose Hamming distance is less than \code{threshold}
(the default is 0.05) are deemed equivalent and assigned to an
equivalence class.}
}
\value{
A list of length the number of equivalence classes, where each
component is a numeric vector of equivalent marker column
numbers. The names of the list are the root marker numbers of the
equivalence classes and can be coerced to \code{numeric}.
}
\description{
Identify redundant genetic genetic markers based on the Hamming
distance between marker pairs.
}
\details{
For a given pair of markers, the Hamming distance is calculated as
the total number of mismatches scaled by n. If \code{na.match =
FALSE}, a missing value in either marker will count as a mismatch
and contribute to the Hamming distance summation.

Beginning with column 1 of \code{M}, the \eqn{m(m-1)/2} pairs of
markers are scanned sequentially and markers allocated to
equivalence classes if the pairwise Hamming distance is less than
\code{threshold}. The algorithm uses the method of D. Eardley in
Section 8.6 of \cite{Press et al., 2002}. Pairwise Hamming
distances are computed on the fly from \code{M} to avoid storing
the distance matrix.
}
\references{
% bibentry: Press:2002
Press WH, Teukolsky SA, Vetterling WT, Flannery BP (2002).
\emph{Numerical Recipes in C++. The Art of Scientific Computing}, Second edition.
Cambridge University Press.
% end:bibentry:  Press:2002
}
