% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all.R
\name{hdist}
\alias{hdist}
\title{Calculate Hamming distances for genetic markers.}
\usage{
hdist(M, na.match = FALSE)
}
\arguments{
\item{M}{An \code{n x m} matrix of marker scores for m markers on n individuals. The
matrix must be numeric and capable of being coerced to integer
values.}

\item{na.match}{If \code{FALSE} (the default), \code{NA} in either marker counts as
a mismatch, that is \code{NAs} match nothing. If \code{TRUE},
\code{NA} in either marker counts as a match, that is \code{NAs}
match everything.}
}
\value{
A \code{dist} class numeric vector of distances being the strict
lower triangle of the distance matrix in column major order.
}
\description{
Calculate the Hamming distance for each marker pair from a matrix
of marker scores.
}
\details{
For a given pair of markers, the Hamming distance is calculated as
the total number of mismatches scaled by n. If \code{na.match =
FALSE}, a missing value in either marker will count as a mismatch
and contribute to the Hamming distance summation.
}
