% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all.R
\name{ld}
\alias{ld}
\title{Linkage disequilibrium}
\usage{
ld(x, method = c("rh", "em"), verbose = FALSE)
}
\arguments{
\item{x}{An \eqn{n \times m} matrix of biallelic marker scores for \eqn{n}
individuals and \eqn{m} markers. The scores are assumed coded as
\{-1, 0, 1\} representing genotypes \code{aa}, \code{aA} and
\code{AA}, respectively, with missing values coded as \code{NA}.}

\item{method}{If method is \code{"rh"} (the default), \eqn{r} is calculated for
each pair of markers in \code{x} using the approximate method of
\cite{Rogers and Huff}, otherwise a modified EM algorithm is used.}

\item{verbose}{If \code{TRUE} and \code{method="em"}, any irregularities in
calculating the haplotype frequencies are reported. Warning: this
may generate considerable console output and would be most useful
in debugging a limited set of markers. The default is \code{FALSE}.}
}
\value{
The \eqn{m(m+1)/2} LD \eqn{r_{ij}} values stored as a dense
symmetric \code{"dspMatrix"} class Matrix; the diagonal elements
are set to 1.
}
\description{
Calculate the LD coefficient \eqn{r} from genotype marker scores.
}
\details{
This function accessess a modification to the EM method of
\cite{Excoffier and Slatkin, 1995}, and the approximation of
\cite{Rogers and Huff, 2009}. The approximation works with pairs of
biallelic loci, and the EM method has been modified to also deal
only with pairs of loci.

The advantages of the approximate method are its ease of coding and
speed improvement over the EM optimisation method, see \cite{Rogers
and Huff, 2009} for details. This implementation uses a simple high
level multithreading strategy for both methods, so timing is
proportional to the number of cores. 

The underlying compiled code is from 
\url{https://github.com/alanrogers/covld} with minor modifications
to capture printed results in \pkg{R}.
}
\references{
% bibentry: Rogers:2009
Rogers AR, Huff C (2009).
``Linkage Disequilibriun Between Loci with Unknown Phase.''
\emph{Genetics}, \bold{182}, 839-844.
% end:bibentry:  Rogers:2009
% bibentry: Excoffier:1995
Excoffier L, Slatkin M (1995).
``Maximum Likelihood Estimation of Molecular Haplotype Frequencies in a Diploid Population.''
\emph{Molecular Biology Evolution}, \bold{12}, 921-927.
% end:bibentry:  Excoffier:1995
}
