% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all.R
\name{ldu}
\alias{ldu}
\title{Generalised Cholesky decomposition.}
\usage{
ldu(A)
}
\arguments{
\item{A}{The numerator relationship matrix.}
}
\value{
A list with the following components:
\describe{
 \item{L}{A lower triangular matrix.}
 \item{D}{A diagonal matrix.}
}
}
\description{
Generalised Cholesky decomposition of a numerator relationship
matrix.
}
\details{
Determines the modified Cholesky decomposition of a
positive-definite matrix \eqn{A} given by \eqn{A=LDL^T}, where
\eqn{L} is lower triangular with ones on the diagonal, and \eqn{D}
is a diagonal matrix. The function uses the \code{MCHOL} function
in \cite{McLeod and Holanda Sales, 1983}.
}
\references{
% bibentry: McLeod:1983
McLeod AI, Holanda
Sales PR (1983).
``Algorithm AS 191.2 An Algorithm for Approximate Likelihood Calculation of ARMA and Seasonal ARMA Models.''
\emph{Applied Statistics}, \bold{32}, 211-223.
% end:bibentry:  McLeod:1983
}
