% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all.R
\name{mask}
\alias{mask}
\title{Mask individual identifiers}
\usage{
mask(pedigree, data = pedigree[, 1], mv = c("0", "*", " "))
}
\arguments{
\item{pedigree}{A data frame with (at least) three columns that correspond to the
individual, male parent and female parent, respectively.}

\item{data}{The names column in data order (with dups); the default is the
first column of the pedigree.}

\item{mv}{missing value indicator.}
}
\value{
A list with components \code{pedigree} and \code{data} containing
the pedigree with renamed individuals and their aliases, respectively.
}
\description{
Replace the identifiers of individuals in the pedigree with
auto generated names using alphabetic combinations.
}
