% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all.R
\name{mat2sparse}
\alias{mat2sparse}
\title{Convert a dense matrix.}
\usage{
mat2sparse(x, rowNames = dimnames(x)[[1]], tol = 1e-09)
}
\arguments{
\item{x}{A dense matrix padded with zeros where needed.}

\item{rowNames}{A character vector set as the "rowNames" attribute of the returned
sparse matrix. The default is the \code{dimnames} attribute of the
leading dimension of \code{x}.}

\item{tol}{Matrix elements whose absolute value is less than tol are considered zero.
The default is 1e-9.}
}
\value{
A matrix holding the lower triangle of a sparse symmetric matrix in
coordinate form in row major order.
}
\description{
Convert a dense matrix to a sparse matrix in three-column
coordinate form.
}
\details{
The \code{dimnames(X)[[1]]} attribute of \code{x} is preserved and returned
as the \code{rowNames} attribute of the returned matrix.
}
