% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all.R
\name{mmat.ev}
\alias{mmat.ev}
\title{Eigenvalues.}
\usage{
mmat.ev(x, k = nrow(x), tol = -1)
}
\arguments{
\item{x}{A real symmetric matrix.}

\item{k}{Calculate the k largest eigenvalues and eigenvectors; the default
is \code{nrow(x)}.}

\item{tol}{Error tolerance to which each eigenvalue is required. If
\code{tol=-1} (the default), it is set to a machine dependent
\emph{safe} minimum.}
}
\value{
\describe{

\item{V}{An \eqn{n \times k} numeric matrix of eigenvectors.}
\item{lambda}{The vector of eigenvalues of length k.}.
}
}
\description{
Calculate the eigenvalues of a symmetric matrix.
}
