% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all.R
\name{mmat}
\alias{mmat}
\title{Realized relationship matrix.}
\usage{
mmat(
  x,
  center = FALSE,
  scale = TRUE,
  na.method = c("mean", "svd", "knn", "none"),
  min.af = NULL,
  max.mv = NULL,
  ev = 10,
  knn = 10,
  maxit = 10,
  imputed = FALSE
)
}
\arguments{
\item{x}{An \ifelse{html}{\out{n x m}}{\eqn{n \times m}} matrix of biallelic marker scores for
\eqn{n} individuals and \eqn{m} markers. The scores are assumed
coded as \{-1, 0, 1\} with missing values coded as \code{NA}.}

\item{center}{If \code{TRUE}, the columns of \code{x} (or the imputed matrix
\eqn{W}) are centred.}

\item{scale}{If \code{TRUE} (the default), the realized association matrix
\eqn{xx'} (or \eqn{WW'}) is scaled by the reciprocal of
the average marker variance (\eqn{v}); that is, \eqn{xx'/v}.
If \code{scale} is numeric then the
realized association matrix is \eqn{xx' * scale}.}

\item{na.method}{\describe{

\item{mean}{missing values within each marker column are replaced
by the column mean;}

\item{knn}{a missing value for a marker is replaced by the
mean of its \code{knn} nearest neighbours with non-missing data
for that individual;}

\item{svd}{iteratively impute missing values in a predicted
(current) \eqn{W_1} using the \code{ev} most significant eigenvalues
from a singular value decomposition of the previous \eqn{W_0};}

\item{none}{no imputation.}
}}

\item{min.af}{Minimum allele frequency below which markers are removed;
monomorphic markers are automatically removed. If not set, the default is
\code{1/(2*nrow(x))}.}

\item{max.mv}{Maximum proportion of missing values allowed for a marker;
completely missing markers are removed. If not set, the default is
\code{1-1/(2*nrow(x))}.}

\item{ev}{If \code{na.method = svd}, the number of significant eigenvectors
to use when imputing missing values (the default is 10).}

\item{knn}{If \code{na.method = knn}, the cluster size of similar markers to
use when imputing missing values (the default is 10).}

\item{maxit}{If \code{na.method = svd}, the maximum number of iterations used to
impute missing values (the default is 10). The iteration sequence may terminate before
\code{maxit} if the convergence criterion is met (see details).}

\item{imputed}{If \code{TRUE} (the default is \code{FALSE}), a list of length two
is returned with components \code{rrm} and \code{inputed}
containing the realized relationship and the imputed marker
matrices, respectively. If \code{FALSE}, only the relationship matrix
is returned.}
}
\value{
A list with components \code{rrm} and \code{imputed}, or just the
relationship matrix object depending on the value
of the \code{imputed} argument. For \code{method=svd}, the
convergence sequence (RSS) is included as attribute \code{rss} of
the returned object.
}
\description{
Calculates an IBS (possibly) additive relationship matrix.
}
\details{
The realized relationship matrix is computed following two
pre-processing steps: 1) the matrix \code{x} is scanned for
consistency against the arguments \code{min.af} and \code{max.mv},
and 2) any missing values are imputed using \code{na.method}.
Monomorphic and completely missing markers are removed in step (1).
If \code{center=TRUE}, the
imputed matrix \eqn{W} is optionally centred, and if
\code{scale=TRUE} \eqn{W} is scaled by the reciprocal of the average marker
variance (\cite{VanRaden 2008}) before computing
\eqn{WW^\prime}. The imputation methods \code{svd} and \code{knn}
replace the missing values in \code{x} with their column means
before proceeding.

The \code{svd} method (\cite{Troyanskaya etal 2001}) uses the
first \code{ev} terms of the SVD of \eqn{W} to replace the
previously missing values with their regression predictions from
the SVD. The procedure iterates until convergence is achieved or
\code{maxit} is exceeded. Convergence is declared when \eqn{|RSS_0
- RSS_1|/RSS_1 < 0.02} (\cite{Rutkoski 2013}), where \eqn{RSS} is
the residual sum of squares between the non-missing values and
their predictions from the SVD, and \eqn{RSS_0} and \eqn{RSS_1} are
the RSS values from successive iterations.

 The \code{knn} procedure (\cite{Troyanskaya etal 2001}) orders the
(marker) neighbours of the \eqn{j^{th}} marker according to
euclidean distance, replacing the \eqn{i^{th}} missing value with
the average of the non-missing values for genotype \eqn{i} from the
\eqn{knn} nearest neighbours of marker \eqn{j}.

If \code{imputed} is \code{TRUE}, the imputed marker matrix is
returned with monomorphic and completely missing markers removed,
as well as those screened out by \code{min.af} and \code{max.mv}.

If \code{scale} is \code{TRUE}, the reciprocal of the average marker
variance is returned in an attribute \code{scale} of the returned
matrix.
}
\examples{

\dontrun{
## Wheat data from Rutkoski (2013) with 70\\% missing
data(WW_VersionNA70_rep1)
## the imputed marker matrix
W <-  mmat(WW_VersionNA70_rep1,na.method="knn",knn=4, imputed=TRUE)
## the realized relationship matrix
WW <-  mmat(WW_VersionNA70_rep1,na.method="knn",knn=4)
}

}
\references{
% bibentry: Rutkoski:2013
Rutkoski JE, Poland J, Jannink J, Sorrells ME (2013).
``Imputation of Unordered Markers and the Impact on Genomic Selection Accuracy.''
\emph{G3 Genes Genomes Genetics}, \bold{3}, 427-439.
% end:bibentry:  Rutkoski:2013
% bibentry: Troyanskaya:2001
Troyanskaya O, Cantor M, Sherlock G, Brown P, Hastie
T. Tibshirani R, Botstein D, Altman RB (2001).
``Missing value estimation methods for DNA microarrays.''
\emph{Bioinformatics}, \bold{17}, 520-525.
% end:bibentry:  Troyanskaya:2001
% bibentry: VanRaden:2008
VanRaden PM (2008).
``Efficient Methods to Compute Genomic Predictions.''
\emph{Journal of Dairy Science}, \bold{91}, 4414-4423.
% end:bibentry:  VanRaden:2008
}
