% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all.R
\name{mmat.svd}
\alias{mmat.svd}
\title{Singular value decomposition.}
\usage{
mmat.svd(x)
}
\arguments{
\item{x}{A real \eqn{m \times n} matrix.}
}
\value{
A list with three components:

\describe{

\item{u}{An \eqn{m \times m} matrix of left singular vectors.}
\item{v}{An \eqn{n \times n} transposed matrix of right singular vectors.}
\item{d}{The \code{min(m,n)} vector of singular values.}
}
}
\description{
Calculate the singular value decomposition of a rectangular matrix.
}
