% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all.R
\name{nrm}
\alias{nrm}
\title{Simulated numerator relationship matrices.}
\usage{
nrm(
  pedigree,
  fgen = list(character(0), 0),
  nsim = 2000,
  rm = c("A", "D", "C", "Ct", "Dh", "Di", "E"),
  form = c("full", "upper"),
  mv = c("NA", "0", "*")
)
}
\arguments{
\item{pedigree}{A data frame with (at least) three columns that correspond to the
individual, male parent and female parent, respectively. The row
giving the pedigree of an individual must appear before any row
where that individual appears as a parent. Founders use 0 (zero) or
\code{NA} in the parental columns.}

\item{fgen}{An optional list of length 2 where \code{fgen[[1]]} is a character
string naming the column in \code{pedigree} that contains the level
of selfing (\eqn{f}) or the level of inbreeding \eqn{b}) of an
individual. In \code{pedigree[,fgen[[1]]]}, 0 indicates a simple
cross, 1 indicates selfed once, 2 indicates selfed twice, etc. A
value between 0 and 1 for a base individual is taken as its
inbreeding value. If the pedigree has implicit individuals (they
appear as parents but not as individuals), they will be assumed
base non-inbred individuals unless their inbreeding level is set
with \code{fgen[[2]]} where \code{0<fgen[[2]]<1} is the inbreeding
level of such individuals. For use in the simulation, \eqn{f} is
determined from \eqn{b} using \eqn{b=1-(0.5)^f} if
\code{0<fgen[[2]]<1}.}

\item{nsim}{The number of traversals of the ancestral tree; default 2000.}

\item{rm}{A character vector naming the relationship matrices to be returned;
the default is to return all simulated matrices. Setting
\code{rm=c("A","D")} will return both the additive and dominance
relationship matrices, for example; see \cite{Lynch and Walsh
(1998)} for details on the components of the genetic covariance
matrix.}

\item{form}{if \code{"full"}, return the complete relationship matrices,
otherwise if \code{form="upper"} the upper triangle is returned.}

\item{mv}{Missing value indicator; elements of \code{pedigree} that exactly
match any element of \code{mv} are treated as missing.}
}
\description{
Calculates components of the genetic covariance matrix for a given
population defined in an ancestral tree.
}
\details{
Initially, each founder individual starts with two distinct (and
unique) alleles.  For each simulation, the pedigree is traversed
and two genes are sampled, one from each parent, and assigned to
each individual. If \eqn{f > 0}, the assigned genes are sampled
with replacement \eqn{f} times, resulting in homozygosity at the
locus if \eqn{f} is sufficiently large. The genes of all
individuals are examined pairwise (\eqn{i,j}), and counts of events
\eqn{d_{ijs}} contributing to identity state S (see \cite{Lynch and
Walsh, 1998}, Chapter 7) are accumulated. If the genes for
individual \eqn{i} are identical, then the count of IBD events
within the \eqn{i^{th}} individual, \eqn{F_i}, is incremented.

The procedure is repeated \code{nsim} (\eqn{N}) times, and the pairwise
coefficients \eqn{D_{ijs}} for state S estimated as \eqn{D_{ijs} =
d_{ijs}/N}, for \eqn{s=1:9}; the inbreeding coefficients are estimated
as \eqn{F_i = F_i/N}. The elements of the various relationship matrices
are estimated using the (pairwise) coefficients of the contributing
identity states as defined in \cite{Lynch and Walsh (1998)}.
}
\references{
% bibentry: Lynch:1998
Lynch M, Walsh B (1998).
\emph{Genetics and Analysis of Quantitative Trails}.
Sinaeur Associates, Sunderland, Massachusetts.
% end:bibentry:  Lynch:1998
% bibentry: Verbyla:2011
Verbyla AP, Butler DG (2011).
``Estimating the Dominance Relationship Matrix Using a Simulation Approach.''
In \emph{The International Biometric Society, Australasian Region Conference}.
% end:bibentry:  Verbyla:2011
}
