% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all.R
\name{offspring}
\alias{offspring}
\title{Count descendents.}
\usage{
offspring(ped)
}
\arguments{
\item{ped}{A data frame with (at least) three columns that correspond to the
individual, male parent and female parent, respectively. The row
giving the pedigree of an individual must appear before any row
where that individual appears as a parent. Founders use 0 (zero) or
\code{NA} in the parental columns.}
}
\value{
A numeric vector of length \code{nrow(ped)} containing the number
of offspring for each individual.
}
\description{
Counts the number of offspring for each individual in a pedigree.
}
