% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all.R
\name{recode.mm}
\alias{recode.mm}
\title{Recode a marker matrix.}
\usage{
recode.mm(
  x,
  map = matrix(c(0, 1, 2, -1, 0, 1), ncol = 2),
  char.convert = TRUE,
  transpose = FALSE
)
}
\arguments{
\item{x}{An \eqn{n\times m} matrix of scores for \eqn{m} molecular markers
on \eqn{n} individuals; missing values are allowed and retained.}

\item{map}{A two column matrix where \code{map[,1]} are the existing scores in
\code{x}, and \code{map[,2]} are the new scores. The default
assumes scores \code{(0,1,2)} are to be recoded as \code{(-1,0,1)}.}

\item{char.convert}{If \code{TRUE} (the default) and \code{x} (and therefore
\code{map}) is of type \code{character}, \code{x} will attempted to
be coerced to \code{numeric} after recoding.}

\item{transpose}{If \code{TRUE} (the default is \code{FALSE}), transpose \code{x} if
rows are markers and columns are individuals.}
}
\value{
The recoded (and possibly transposed) matrix with missing values
preserved.
}
\description{
Recode the matrix of molecular marker scores using a simple
mapping.
}
