% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all.R
\name{sparse2mat}
\alias{sparse2mat}
\title{Convert a sparse matrix.}
\usage{
sparse2mat(x)
}
\arguments{
\item{x}{A matrix (or data frame) holding the lower triangle of a sparse
symmetric matrix in coordinate form in row major order; typically
the result of a call to \code{ainv}.}
}
\value{
A dense matrix containing \code{x} padded with zeros.
}
\description{
Convert a sparse matrix held in three-column coordinate form to a
dense matrix.
}
\details{
The attribute \code{rowNames} of \code{x} is preserved and returned
as the \code{dimnames} attribute of the returned matrix.
}
