% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all.R
\name{trim.par}
\alias{trim.par}
\title{Extract parental records from a (pruned) pedigree.}
\usage{
trim.par(ped)
}
\arguments{
\item{ped}{A data frame with (at least) three columns that correspond to the
individual, male parent and female parent, respectively; typically
the result of a previous call to \code{\link{trim}}.}
}
\value{
A data frame with the same structure as \code{ped}, with the pruned
individuals removed.
}
\description{
Trim a pedigree to the set of unique identifiers in the parental
columns.
}
