% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all.R
\name{trim}
\alias{trim}
\title{Trim a pedigree.}
\usage{
trim(ped, data, gen = NULL)
}
\arguments{
\item{ped}{A data frame with (at least) three columns that correspond to the
individual, male parent and female parent, respectively. The row
giving the pedigree of an individual must appear before any row
where that individual appears as a parent. Founders use 0 (zero) or
\code{NA} in the parental columns.}

\item{data}{A data frame with a component named \code{names(ped)[1]} containing
the identities of the individuals with data.}

\item{gen}{The number of generations to keep before those with data. The default
is \code{max(count_gen(mmd))}.}
}
\value{
A data frame with the same structure as \code{ped}, with the pruned
individuals removed.
}
\description{
Trim a pedigree to a given data frame; non-informative individuals
without data are removed.
}
\details{
Call \code{chkPed} (library \code{asreml}) to add any missing
founders, resolve \code{fgen}, and reorder the pedigree (if
necessary) prior to pruning.
}
